fix prod build issue with schemas
