+++ app/_components/loginactions.tsx
@@ -3,7 +3,7 @@ import { createClient } from '@/utils/supabase/server'
 import { redirect } from 'next/navigation';
 import { headers, cookies } from 'next/headers'
 import { formSchema } from '../post/page';
-import { signinFormSchema, signupFormSchema } from '../login/page' 
+import { signinFormSchema, signupFormSchema } from './schemas' 
 import { z } from 'zod';
