+++ app/login/page.tsx
@@ -16,56 +16,9 @@ import {
   FormMessage,
 } from "../_components/ui/form"
 import { Input } from "../_components/ui/input"
+import { signinFormSchema, signupFormSchema } from '../_components/schemas'
 
-export const signinFormSchema = z.object({
-  email: z.string().email({
-      message: 'Please enter a valid email address'
-  }),
-  password: z.string().min(8, {
-      message: 'Password must be at least 8 characters long'
-  }),
-})
 
-export const signupFormSchema = z.object({
-  username: z.string().min(2, {
-      message: 'Username must be at least 2 characters long'
-  }),
-  email: z.string().email({
-      message: 'Please enter a valid email address'
-  }),
-  password: z.string().min(8, {
-      message: 'Password must be at least 8 characters long'
-  })
-
-})
-// .superRefine(({ password }, checkPassComplexity) => {
-//   const containsUppercase = (ch: string) => /[A-Z]/.test(ch);
-//   const containsLowercase = (ch: string) => /[a-z]/.test(ch);
-//   const containsSpecialChar = (ch: string) =>
-//     /[`!@#$%^&*()_\-+=\[\]{};':"\\|,.<>\/?~ ]/.test(ch);
-//   let countOfUpperCase = 0,
-//     countOfLowerCase = 0,
-//     countOfNumbers = 0,
-//     countOfSpecialChar = 0;
-//   for (let i = 0; i < password.length; i++) {
-//     let ch = password.charAt(i);
-//     if (!isNaN(+ch)) countOfNumbers++;
-//     else if (containsUppercase(ch)) countOfUpperCase++;
-//     else if (containsLowercase(ch)) countOfLowerCase++;
-//     else if (containsSpecialChar(ch)) countOfSpecialChar++;
-//   }
-//   if (
-//     countOfLowerCase < 1 ||
-//     countOfUpperCase < 1 ||
-//     countOfSpecialChar < 1 ||
-//     countOfNumbers < 1
-//   ) {
-//     checkPassComplexity.addIssue({
-//       code: "custom",
-//       message: "password must have at least 1 uppercase, 1 lowercase, 1 number, and 1 special character",
-//     });
-//   }
-// })
 
 export default function Login({
   searchParams,
